Move Git* classes to FileFinders because it is only used there
